requirements.txt
boto3==1.26.122
botocore==1.29.122
celery==5.3.1
click==8.1.3
Flask==2.3.1
flask_sqlalchemy==3.0.3
flask_cors==4.0.0
langchain==0.0.352
openai==0.27.8
python-dotenv==1.0.0
Werkzeug==2.3.1
invoke==2.2.0
pinecone-client==3.2.2
pypdf==3.15.4
watchdog==3.0.0
tiktoken==0.4.0
uuid==1.30
langfuse==1.0.18
backoff==2.2.1
redis==5.0.0
langchain_pinecone
langchain-core~=0.1.23
requests~=2.32.5
SQLAlchemy~=2.0.44
langchain-openai~=0.0.5
pydantic~=1.10.26